% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knlib_data_functions.R
\name{kn_data_search}
\alias{kn_data_search}
\title{Search dataset in KN}
\usage{
kn_data_search(q, offset, limit)
}
\arguments{
\item{q}{Search query}

\item{offset}{Search offset}

\item{limit}{Number of results to return}
}
\value{
Data frame result as <pid, title> from KN
}
\description{
Search dataset in KN
}
\examples{
kn_data_search('water', 0, 5)
}
