% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knlib_data_functions.R
\name{kn_get_geojson}
\alias{kn_get_geojson}
\alias{kn_get_geojson}
\title{Get GeoJSON from KN}
\usage{
kn_get_geojson(pid)

kn_get_geojson(pid)
}
\arguments{
\item{pid}{KN Spatial Identifier}

\item{asSP}{Return as spatial object (TRUE, default) or plain geojson (FALSE)}

\item{pid}{KN Spatial Identifier}
}
\value{
Geojson object from KN

Geojson object from KN
}
\description{
Get GeoJSON from KN

Get GeoJSON from KN
}
\examples{
kn_get_geojson('http://oznome.csiro.au/id/geo/geofabric/9400238')
kn_get_geojson('http://oznome.csiro.au/id/geo/geofabric/9400238')
}
