% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knlib_data_functions.R
\name{knsearch_list}
\alias{knsearch_list}
\title{Search dataset in KN and return a list}
\usage{
knsearch_list(q, type, offset, limit)
}
\arguments{
\item{q}{Search query}

\item{type}{Type of thing to search for - either "data" or "feature"}

\item{offset}{Search offset}

\item{limit}{Number of results to return}
}
\value{
Result list from KN with all fields in each item
}
\description{
Search dataset in KN and return a list
}
\examples{
knsearch_list('water', 'data', 0, 5)
}
